Add inheritance to Purchase class
