+++ src/main/java/seedu/jarvis/model/financeTracker/Purchase.java
@@ -3,15 +3,10 @@ package seedu.jarvis.model.financeTracker;
 public class Purchase {
     private String description;
     private double moneySpent;
-    private PersonPaid personPaid = null;
 
     public Purchase(String description, double moneySpent) {
         this.description =  description;
         this.moneySpent = moneySpent;
-    }
-
-    public void setPersonPaid(PersonPaid p) {
-        this.personPaid = p;
     }
 
     /**
@@ -25,18 +20,8 @@ public class Purchase {
         return this.moneySpent;
     }
 
-    public PersonPaid getPersonPaid() {
-        return this.personPaid;
-    }
-
     @Override
     public String toString() {
-        String str = "";
-        if (personPaid == null) {
-            str = this.description + " (" + this.moneySpent + ")";
-        } else {
-            str = this.description + " paid to " + this.personPaid + " (" + this.moneySpent + ")";
-        }
-        return str;
+        return this.description + " (" + this.moneySpent + ")";
     }
 }
